try to solve problem of ent 0

+++ gailtf/baselines/common/distributions.py
@@ -188,7 +188,7 @@ class CategoricalPd(Pd):
         def f1(): return tf.reshape(tf.multinomial(available_logits, 1), [])
         def f2(): 
             # random.seed(datetime.now())
-            return tf.reshape(tf.random_uniform([], maxval=tf.count_nonzero(available_action, dtype=tf.float32)-1), [])
+            return tf.cast(tf.reshape(tf.random_uniform([], maxval=tf.count_nonzero(available_action, dtype=tf.float32)-1), []), tf.int32)
 
         ent = tf.reshape(self.entropy(), [])
         index_for_available_act = tf.cond(
